create pagecustom resolve problem count element

+++ src/main/java/com/nexigroup/pagopa/cruscotto/sert/service/dto/WorkflowResponseDTO.java
@@ -16,7 +16,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @NoArgsConstructor
 @AllArgsConstructor
 public class WorkflowResponseDTO implements Serializable {
-    private Double count;
+    private Long count;
 
     @JsonProperty("events-position")
     private List<WorkflowObjectDTO> eventsPosition;
